feat: adicionar hook para obter lista de alimentos paginada e refatorar modelo de alimentos
