+++ src/pages/Foods/MyFoods/my-foods-model.ts
@@ -1,8 +1,7 @@
 import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useInfiniteQuery } from '@tanstack/react-query'
+import { useGetPaginatedFoodList } from '@/hooks/use-get-paginated-food-list'
 import { IGetPaginatedFoodListService } from '@/services/http/food/food-service'
-import { QueryKeys } from '@/types/query-keys'
 
 type UseMyFoodsModelProps = {
   getPaginatedFoodListService: IGetPaginatedFoodListService
@@ -15,35 +14,8 @@ export function useMyFoodsModel({
 
   const [searchTerm, setSearchTerm] = useState('')
 
-  const { data, isPending, isFetchingNextPage, hasNextPage, fetchNextPage } =
-    useInfiniteQuery({
-      queryKey: [QueryKeys.FoodList],
-      queryFn: async ({ pageParam }) => {
-        const response = await getPaginatedFoodListService({
-          page: pageParam,
-          pageSize: 10,
-        })
-
-        return response
-      },
-      initialPageParam: 1,
-      getNextPageParam: lastPage => {
-        if (lastPage.currentPage + 1 > lastPage.totalPages) {
-          return undefined
-        }
-
-        return lastPage.currentPage + 1
-      },
-      staleTime: 15 * 60 * 1000,
-    })
-
-  const handlePagination = () => {
-    if (!isFetchingNextPage && hasNextPage) {
-      fetchNextPage()
-    }
-  }
-
-  const foods = data?.pages.flatMap(page => page.items) || []
+  const { foods, isFoodsLoading, handlePagination, hasNextPage } =
+    useGetPaginatedFoodList({ getPaginatedFoodListService })
 
   const filteredFoods = foods?.filter(food =>
     food.name.toLowerCase().includes(searchTerm.toLowerCase()),
@@ -54,7 +26,7 @@ export function useMyFoodsModel({
 
   return {
     foods: filteredFoods || [],
-    isFoodsLoading: isPending,
+    isFoodsLoading,
     handlePagination,
     hasNextPage,
     searchTerm,
